feat(backend/runner): the shared attachment pipeline extracts archives and honours mountPath inside the inputs namespace
S3 M1 commit 7 (Q-S3-14, Q-M1-2). The #567 ZIP guards (central-directory
validation, traversal, zip bombs, declared-size re-enforcement) move
verbatim from the native injector into shared/attachment-zip.ts, one
home; the injector imports them until M2b retires it, and the 20 guard
tests move with the code.

shared/attachment-resolver.ts, the runtime's one attachment pipeline,
gains what the injector had and it lacked: an extract attachment lands
under its mount as a directory (inputs/<mount>/<entry>), one resolved
attachment per entry with no vision, rename or URL; an explicit mountPath
is honoured INSIDE the inputs/ namespace (the CLI's directory
attachments, the console's approved plan) in a two-pass placement that
refuses two explicit claims on one path, and REFUSED outside it with an
actionable error — a file in the user's tree would ride the write-back's
git add -A into the session PR (F-M1-2b). Cursor gains both as an
intended alignment; 22 + 17 goldens byte-identical.
